core/requirements.txt
anyio==4.9.0
asgiref==3.9.1
attrs==25.3.0
build==1.2.2.post1
CacheControl==0.14.3
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
cleo==2.1.0
crashtest==0.4.1
distlib==0.4.0
Django==5.2.4
django-cors-headers==4.7.0
django-countries==7.6.1
django-jazzmin @ git+https://github.com/farridav/django-jazzmin.git@d269874c96fe5ff6d231c3017cdd4650bc2f3105
djangorestframework==3.16.0
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.7.1
drf-yasg==1.21.10
dulwich==0.21.7
fastjsonschema==2.21.1
filelock==3.18.0
findpython==0.6.3
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
inflection==0.5.1
installer==0.7.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.2.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
keyring==24.3.1
more-itertools==10.7.0
msgpack==1.1.1
packaging==25.0
pbs-installer==2025.7.23
pexpect==4.9.0
pillow==11.3.0
pkginfo==1.12.1.2
platformdirs==4.3.8
poetry-core==1.9.1
poetry-plugin-export==1.8.0
ptyprocess==0.7.0
pycparser==2.22
pyproject_hooks==1.2.0
pytz==2025.2
PyYAML==6.0.2
RapidFuzz==3.13.0
referencing==0.36.2
requests==2.32.4
requests-toolbelt==1.0.0
rpds-py==0.26.0
shellingham==1.5.4
sniffio==1.3.1
sqlparse==0.5.3
tomlkit==0.13.3
trove-classifiers==2025.5.9.12
typing_extensions==4.14.1
uritemplate==4.2.0
urllib3==2.5.0
virtualenv==20.32.0
xattr==1.2.0
zstandard==0.23.0
